Moved tests to dualview file
